Reorder the termination checks
The previous order of the termination checks occasionally allowed the run to go past the maximum number of generations. This moves that check up higher so that we are guaranteed to terminate when we run out of generations.

We also checked that we were not using quick-check before terminating when the total error is zero.