Fix the ERC settings
The GECCO paper says the only constant is 0, where the Clojush implementation actually provides random integers in the range [-50, 50]. I talked to Tom Helmuth, and it's clearly just an "oops" that these aren't the same. We should probably use the Clojush version for better comparisons, which is what I've done here.

We might also want to include 0 and see what impact that has, as it seems likely to make the problem easier.