Beta version of 'last-index-of-zero'
Adds `error-function` and  training data to the 'last-index-of-zero' problem. The test data is empty at the moment and will need to eventually be populated. We also renamed the file so that it is compatible with ClojureScript.

The training data currently comprises of the edge case data from https://github.com/thelmuth/program-synthesis-benchmark-datasets/tree/master/datasets/last-index-of-zero

We should also eventually bring in the random cases as test data.

We believe this should work, however we haven't been able to do a run due to problems with some of the vector instructions.