Fix `indexof` and updated `_replacefirst`
`indexof` did not work if stack type was boolean, so we replaced it with a new version that should fix this problem.

`_replacefirst` did not check for the case that `indexof` does not find the element and so returns -1, so we added an if statement to check for said case.