Add support for custom result equality checks in quick-check
We may not want to always use `=` to decide if a program's output matches the target output when generating new test cases in Quick Check. This looks in `argmap` for a `:result-equality-test` key, and defaults to `=` if there is none.

We also removed the call to `read-string` since that only works if the program's result is a string and the target value is something else (like a number). We think it makes more sense to change the test cases so the expected outputs have the same type as the value returned by the program.