Change type of expected output for printing problems to strings
Both smallest and number-io print their results (i.e., return strings), but the code stored the expected outputs as numbers, which created issues in quick-check. @ErikRauer and I decided to change the outputs to match their actual types. This involved calling `str` in `target-function` to convert the numeric result to a string, and parsing all those in `error-function` so the difference computations will work there.