Add `:exec` instructions to Last Index of Zero
We'd forgotten `:exec`, so there were no looping instructions, and basically no chance of solving the problem in a general form.

I also needed to add `'close` so evolution could "learn" when to close blocks. I found I needed to add quite a few of them to get any kind of useful structure, but 20 is an annoyingly arbitrary number (& I don't even know if it's a _good_ number).