Adapt to new data file formats (sort of)
There's been change in the file formats on https://github.com/thelmuth/program-synthesis-benchmark-datasets, and we needed to modify the code in `last_index_of_zero.cljc` to read in the new format.

What we have here cheats in one significant way – I hacking a pair of open and close square brackets (`[` and `]`) at the beginning and end of both data files so we could read the whole dataset in as a single operation. That's not the right way to do it, though, and we'll want to do that "right".

Co-authored-by: rauer007@morris.umn.edu