Add `checksum` problem
This is based on the implementation of checksum in Clojush.

Similar to `last-index-of-zero` this requires the edn files from https://github.com/thelmuth/program-synthesis-benchmark-datasets/tree/master/datasets/checksum. It also currently requires that you add a set of square brackets around the full set of cases in both edn files.